refactor(phase5): tidy shell + repo root
- contexts/WatchlistContext -> app/providers/WatchlistContext (+ fix import & vi.mock)
- remove empty src/app/dev and empty root workflows/
- declutter root: find-unused-files.js + check-status.sh -> scripts/maintenance/;
  PLANNING.md + consistency-audit.md -> docs/
Left scripts/ internal one-offs in place (contained; relocating risks ./utils +
.sh-wrapper coupling for marginal gain). No behavior change (8/8 e2e green).
