+++ scripts/maintenance/check-status.sh
