+++ src/App.jsx
@@ -3,7 +3,7 @@ import * as Sentry from '@sentry/react'
 import { RouterProvider } from 'react-router-dom'
 import { QueryClientProvider } from '@tanstack/react-query'
 import { router } from '@/app/router'
-import { WatchlistProvider } from '@/contexts/WatchlistContext'
+import { WatchlistProvider } from '@/app/providers/WatchlistContext'
 import { useSessionTracking } from '@/shared/hooks/useInteractionTracking'
 import { queryClient } from '@/shared/queryClient'
