+++ src/components/carousel/CardContent/MovieCard.jsx
@@ -2,7 +2,7 @@ import { memo, useCallback, useMemo, useRef, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
 import { Check, Eye, EyeOff, Plus } from 'lucide-react'
 import { tmdbImg, posterSrcSet } from '@/shared/api/tmdb'
-import { useWatchlistContext } from '@/contexts/WatchlistContext'
+import { useWatchlistContext } from '@/app/providers/WatchlistContext'
 import { useUserMovieStatus } from '@/shared/hooks/useUserMovieStatus'
 import { updateImpression } from '@/shared/services/recommendations'
 import { track } from '@/shared/services/analytics'
